Improve performance for transitive dependency checks
`TransitiveDependencyCondition` internally calls `contains()` recursively
on the collection of all objects to be tested. If this collection is a
large list and there are enough recursive calls to
`getDirectDependencyTargetsOutsideOfAnalyzedClasses()` this results in a
heavy performance impact. On a reasonable large project a single test
using that condition may take minutes to complete. Converting the given
list to a Set with much better `contains()` performance fixes this issue.

on-behalf-of: @e-solutions-GmbH <info@esolutions.de>
Signed-off-by: To6i <To6i@users.noreply.github.com>